+++ java/daikon/FileIO.java
@@ -1016,6 +1016,9 @@ public final class FileIO {
    * Reads data from {@code .dtrace} files. For each record in the files, calls the appropriate
    * callback in the processor.
    *
+   * @param files the list of {@code .dtrace} files to read. If {@link Daikon#server_dir} is
+   *     non-null, this list may be appended to.
+   * @param all_ppts the set of all program points; may be increased by this method
    * @see #read_data_trace_files(Collection,PptMap,Processor,boolean)
    * @see #read_data_trace_file(String,PptMap,Processor,boolean,boolean)
    */
@@ -1030,6 +1033,10 @@ public final class FileIO {
    * Reads data from {@code .dtrace} files. Calls {@link
    * #read_data_trace_file(String,PptMap,Processor,boolean,boolean)} for each element of filenames.
    *
+   * @param files the list of {@code .dtrace} files to read. If {@link Daikon#server_dir} is
+   *     non-null, this list may be appended to.
+   * @param all_ppts the set of all program points; may be increased by this method
+   * @param processor the function that processes dtrace entries
    * @param ppts_may_be_new true if declarations of ppts read from the data trace file are new (and
    *     thus are not in all_ppts). false if the ppts may already be there.
    * @see #read_data_trace_file(String,PptMap,Processor,boolean,boolean)
@@ -1885,7 +1892,7 @@ public final class FileIO {
       // Rather than defining leaves as :::EXIT54 (numbered exit)
       // program points define them as everything except
       // ::EXIT (combined), :::ENTER, :::THROWS, :::OBJECT, ::GLOBAL
-      //  and :::CLASS program points.  This scheme ensures that arbitrarly
+      //  and :::CLASS program points.  This scheme ensures that arbitrarily
       //  named program points such as :::POINT (used by convertcsv.pl)
       //  will be treated as leaves.
 
@@ -2950,7 +2957,7 @@ public final class FileIO {
       // System.out.printf("flags for %s are %s%n", name, flags);
     }
 
-    /** Parse the langauge specific flags record. Multiple flags can be specified. */
+    /** Parse the language-specific flags record. Multiple flags can be specified. */
     public void parse_lang_flags(Scanner scanner) {
 
       lang_flags.add(parse_enum_val(scanner, LangFlags.class, "Language Specific Flag"));