+++ java/daikon/suppress/NISuppressor.java
@@ -181,8 +181,8 @@ public class NISuppressor {
 
   /**
    * Returns true if this suppressor invariant could be instantiated over the specified variables. A
-   * suppressor that canot be instantiated over the variables cannot possibly suppress. Consider the
-   * NonZero invariant. It is suppressed by EqualsOne. But while NonZero is valid over all
+   * suppressor that cannot be instantiated over the variables cannot possibly suppress. Consider
+   * the NonZero invariant. It is suppressed by EqualsOne. But while NonZero is valid over all
    * variables, EqualsOne is only valid over non-pointer variables. Thus the suppression is only
    * valid over non-pointer variables.
    */