+++ java/daikon/test/diff/DiffDummyInvariant.java
@@ -22,7 +22,7 @@ public class DiffDummyInvariant extends Invariant {
   /** A string representation of the formula that this dummy invariant represents. */
   public String formula;
 
-  /** The confidence fro this dummy invariant. */
+  /** The confidence for this dummy invariant. */
   public double confidence;
 
   /** Whether the invariant is worth printing. */