+++ java/daikon/tools/DtraceNonceFixer.java
@@ -48,7 +48,7 @@ public class DtraceNonceFixer {
 
   /**
    * This does the work of {@link #main(String[])}, but it never calls System.exit, so it is
-   * appropriate to be called progrmmatically.
+   * appropriate to be called programmatically.
    *
    * @param args command-line arguments, like those of {@link #main}
    */